backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
google-generativeai>=0.8.0
chromadb>=0.5.0
python-dotenv>=1.0.0
supabase>=2.0.0
prometheus-client>=0.21.0
